chore: Refactor the Link model
